gamestate time refactor
