+++ src/plugins/PhysicsCharacterController.ts
@@ -183,7 +183,7 @@ export const PlayerControllerSystem = (state: GameState) => {
       shapeCastRotation,
       physicsWorld.gravity,
       colliderShape,
-      state.time.dt,
+      state.dt,
       CharacterShapecastInteractionGroup
     );
 
@@ -198,7 +198,7 @@ export const PlayerControllerSystem = (state: GameState) => {
         .set(moveVec[0], 0, -moveVec[1])
         .normalize()
         .applyQuaternion(obj.quaternion)
-        .multiplyScalar(walkSpeed * state.time.dt);
+        .multiplyScalar(walkSpeed * state.dt);
 
       if (!isGrounded) {
         moveForce.multiplyScalar(inAirModifier);
@@ -215,13 +215,13 @@ export const PlayerControllerSystem = (state: GameState) => {
       speed > minSlideSpeed &&
       isGrounded &&
       !isSliding &&
-      state.time.dt > lastSlideTime + slideCooldown
+      state.dt > lastSlideTime + slideCooldown
     ) {
       slideForce.set(0, 0, (speed + 1) * -slideModifier).applyQuaternion(obj.quaternion);
       moveForce.add(slideForce);
       isSliding = true;
-      lastSlideTime = state.time.elapsed;
-    } else if (crouch.released || state.time.dt > lastSlideTime + slideCooldown) {
+      lastSlideTime = state.elapsed;
+    } else if (crouch.released || state.dt > lastSlideTime + slideCooldown) {
       isSliding = false;
     }
 
@@ -237,7 +237,7 @@ export const PlayerControllerSystem = (state: GameState) => {
       dragForce
         .copy(linearVelocity)
         .negate()
-        .multiplyScalar(dragMultiplier * state.time.dt);
+        .multiplyScalar(dragMultiplier * state.dt);
 
       dragForce.y = 0;
